A-MoreOOP: Modified Ui (deleteTask, printDeleteEvent, doneTask, printEvent(Todo, Deadline, Event's tasks)
